[sprint_3] added ExceptionHandler for Author

+++ src/main/java/kz/halykacademy/bookstore/services/impl/AuthorServiceImpl.java
@@ -2,89 +2,84 @@ package kz.halykacademy.bookstore.services.impl;
 
 import kz.halykacademy.bookstore.dto.AuthorDTO;
 import kz.halykacademy.bookstore.dto.GenreDTO;
+import kz.halykacademy.bookstore.errors.ClientBadRequestException;
+import kz.halykacademy.bookstore.dto.ModelResponseDTO;
+import kz.halykacademy.bookstore.errors.ResourceNotFoundException;
 import kz.halykacademy.bookstore.models.Author;
 import kz.halykacademy.bookstore.repositories.AuthorRepository;
 import kz.halykacademy.bookstore.services.AuthorService;
-import org.modelmapper.ModelMapper;
+import kz.halykacademy.bookstore.utils.authorconvertor.AuthorConvertor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.Assert;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static kz.halykacademy.bookstore.utils.AssertUtil.notNull;
+
 @Service
 @Transactional
 public class AuthorServiceImpl implements AuthorService {
+    private final String MESSAGE_NOT_FOUND = "Author is not found with id = %d";
+    private final String MESSAGE_SUCCESS = "success";
+    private final String MESSAGE_EXISTED = "This author is existed";
+
     private final AuthorRepository authorRepository;
-    private final ModelMapper modelMapper;
+    private final AuthorConvertor authorConvertor;
 
     @Autowired
-    public AuthorServiceImpl(AuthorRepository authorRepository, ModelMapper modelMapper) {
+    public AuthorServiceImpl(AuthorRepository authorRepository, AuthorConvertor authorConvertor) {
         this.authorRepository = authorRepository;
-        this.modelMapper = modelMapper;
+        this.authorConvertor = authorConvertor;
     }
 
     @Override
     public ResponseEntity create(AuthorDTO authorDTO) {
-        String message;
+        // проверка параметров запроса
+        checkParameters(authorDTO);
 
-        try {
-            // проверка параметров запроса
-            checkParameters(authorDTO);
+        // конвертирование DTO в Entity
+        Author author = authorConvertor.convertToAuthor(authorDTO);
 
-            // конвертирование DTO в Entity
-            Author author = convertToAuthor(authorDTO);
+        // Поиск автора в БД
+        Author foundAuthor = authorRepository
+                .findAuthorByNameAndSurnameAndLastname(
+                        author.getName(),
+                        author.getSurname(),
+                        author.getLastname());
 
-            // Поиск автора в БД
-            Author foundAuthor = authorRepository
-                    .findAuthorByNameAndSurnameAndLastname(
-                            author.getName(),
-                            author.getSurname(),
-                            author.getLastname());
+        // Проверка существует ли автор
+        if (foundAuthor == null) {
+            // если нет, то создаем
+            authorRepository.save(author);
 
-            // Проверка существует ли автор
-            if (foundAuthor == null) {
-                // если нет, то создаем
-                authorRepository.save(author);
-                message = "success";
-                return new ResponseEntity(message, HttpStatus.OK);
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(new ModelResponseDTO(String.format(MESSAGE_SUCCESS)));
 
-            } else {
-                // иначе выводим сообщение пользователю
-                message = "This author is existed";
-                return new ResponseEntity(message, HttpStatus.BAD_REQUEST);
-            }
-
-        } catch (Exception e) {
-            message = "Internal Server Error: " + e.getMessage();
-            return new ResponseEntity(message, HttpStatus.INTERNAL_SERVER_ERROR);
+        } else {
+            // иначе выводим сообщение пользователю
+            throw new ClientBadRequestException(String.format(MESSAGE_EXISTED));
         }
     }
 
     @Override
     public ResponseEntity readById(Long id) {
-        try {
-            // Поиск автора по id
-            Optional<Author> authorById = authorRepository.findById(id);
+        // Поиск автора по id
+        Optional<Author> authorById = authorRepository.findById(id);
 
-            if (authorById.isEmpty()) {
-                // Если не найден автор
-                return new ResponseEntity("Author is not found", HttpStatus.BAD_REQUEST);
-            }
-
-            AuthorDTO authorDTO = convertToAuthorDTO(authorById.get());
-            return new ResponseEntity(authorDTO, HttpStatus.OK);
-
-        } catch (Exception e) {
-            String message = "Internal Server Error: " + e.getMessage();
-            return new ResponseEntity(message, HttpStatus.INTERNAL_SERVER_ERROR);
+        if (authorById.isEmpty()) {
+            // Если не найден автор
+            throw new ResourceNotFoundException(String.format(MESSAGE_NOT_FOUND, id));
         }
+
+        AuthorDTO authorDTO = authorConvertor.convertToAuthorDTO(authorById.get());
+
+        return new ResponseEntity(authorDTO, HttpStatus.OK);
     }
 
     @Override
@@ -92,50 +87,51 @@ public class AuthorServiceImpl implements AuthorService {
         return authorRepository
                 .findAll()
                 .stream()
-                .map(this::convertToAuthorDTO)
+                .map(authorConvertor::convertToAuthorDTO)
                 .collect(Collectors.toList());
     }
 
     @Override
     public ResponseEntity update(Long id, AuthorDTO updatedAuthorDTO) {
-        String message;
 
-        try {
-            // проверка параметров запроса
-            checkParameters(id, updatedAuthorDTO);
+        // проверка параметров запроса
+        checkParameters(id, updatedAuthorDTO);
 
-            // Поиск автора по id
-            Optional<Author> author = authorRepository.findById(id);
-            if (author.isPresent()) {
-                // Если найден, обновляем автора
-                authorRepository.save(convertToAuthor(updatedAuthorDTO));
-                message = "successfully updated";
-                return new ResponseEntity(message, HttpStatus.OK);
+        // Поиск автора по id
+        Optional<Author> author = authorRepository.findById(id);
 
-            } else {
-                // иначе выводим сообщение пользователю
-                message = "Author is not found";
-                return new ResponseEntity(message, HttpStatus.BAD_REQUEST);
-            }
+        if (author.isPresent()) {
+            // Если найден, обновляем автора
+            authorRepository.save(authorConvertor.convertToAuthor(updatedAuthorDTO));
 
-        } catch (Exception e) {
-            message = "Internal Server Error: " + e.getMessage();
-            return new ResponseEntity(message, HttpStatus.INTERNAL_SERVER_ERROR);
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(new ModelResponseDTO(String.format(MESSAGE_SUCCESS)));
+
+        } else {
+            // иначе выводим сообщение пользователю
+            throw new ResourceNotFoundException(String.format(MESSAGE_NOT_FOUND, id));
         }
     }
 
     @Override
     public ResponseEntity delete(Long id) {
-        try {
-            Assert.notNull(id, "Id is undefined");
+        // Проверка параметра id
+        notNull(id, "Id is undefined");
 
+        // Поиск автора по Id
+        Optional<Author> foundAuthor = authorRepository.findById(id);
+
+        if (foundAuthor.isPresent()) {
+            // если нашли, удаляем
             authorRepository.deleteById(id);
 
-            return new ResponseEntity("success", HttpStatus.OK);
-
-        } catch (Exception e) {
-            String message = "Internal Server Error: " + e.getMessage();
-            return new ResponseEntity(message, HttpStatus.INTERNAL_SERVER_ERROR);
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(new ModelResponseDTO(String.format(MESSAGE_SUCCESS)));
+        } else {
+            // иначе выводим сообщение пользователю
+            throw new ResourceNotFoundException(String.format(MESSAGE_NOT_FOUND, id));
         }
     }
 
@@ -145,7 +141,7 @@ public class AuthorServiceImpl implements AuthorService {
         return authorRepository
                 .findByNameOrSurnameOrLastnameLike("", "", "")
                 .stream()
-                .map(this::convertToAuthorDTO)
+                .map(authorConvertor::convertToAuthorDTO)
                 .collect(Collectors.toList());
     }
 
@@ -156,29 +152,14 @@ public class AuthorServiceImpl implements AuthorService {
         return null;
     }
 
-    private AuthorDTO convertToAuthorDTO(Author author) {
-        return modelMapper.map(author, AuthorDTO.class);
-    }
-
-    private Author convertToAuthor(AuthorDTO authorDTO) {
-        return modelMapper.map(authorDTO, Author.class);
-    }
-
     protected void checkParameters(Long id, AuthorDTO authorDTO) {
-        Assert.notNull(id, "Id is undefined");
+        notNull(id, "Id is undefined");
         checkParameters(authorDTO);
     }
 
     protected void checkParameters(AuthorDTO authorDTO) {
         notNull(authorDTO.getName(), "Name is undefined");
         notNull(authorDTO.getSurname(), "Surname is undefined");
-        Assert.notNull(authorDTO.getDateOfBirth(), "Date of birth is undefined");
+        notNull(authorDTO.getDateOfBirth(), "Date of birth is undefined");
     }
-
-    public static void notNull(@Nullable String object, String message) {
-        if (object == null || object.equals("")) {
-            throw new IllegalArgumentException(message);
-        }
-    }
-
 }
